refactor: rename command to commands
