+++ src/joshuto/commands/change_directory.rs
@@ -3,7 +3,7 @@ extern crate ncurses;
 use std::path;
 use std::process;
 
-use joshuto::command::{JoshutoCommand, JoshutoRunnable};
+use joshuto::commands::{JoshutoCommand, JoshutoRunnable};
 use joshuto::context::JoshutoContext;
 use joshuto::preview;
 use joshuto::ui;
