+++ src/joshuto/commands/set_mode.rs
@@ -1,6 +1,6 @@
 extern crate ncurses;
 
-use joshuto::command::{JoshutoCommand, JoshutoRunnable};
+use joshuto::commands::{JoshutoCommand, JoshutoRunnable};
 use joshuto::context::JoshutoContext;
 use joshuto::structs::JoshutoDirEntry;
 use joshuto::textfield::JoshutoTextField;
