+++ src/joshuto/commands/tab_switch.rs
@@ -1,4 +1,4 @@
-use joshuto::command::{JoshutoCommand, JoshutoRunnable};
+use joshuto::commands::{JoshutoCommand, JoshutoRunnable};
 use joshuto::context::JoshutoContext;
 use joshuto::ui;
